+++ Rdmp.Core/DataLoad/Modules/DataFlowSources/DelimitedFlatFileDataFlowSource.cs
@@ -393,7 +393,7 @@ This will not help you avoid bad data as the full file structure must still be r
                 HasHeaderRecord = string.IsNullOrWhiteSpace(ForceHeaders),
                 ShouldSkipRecord = ShouldSkipRecord,
                 IgnoreBlankLines = IgnoreBlankLines,
-                Mode = IgnoreQuotes ? CsvMode.Escape : CsvMode.RFC4180,
+                Mode = IgnoreQuotes ? CsvMode.NoEscape : CsvMode.RFC4180,
                 BadDataFound = s => EventHandlers.BadDataFound(new FlatFileLine(s), true),
                 ReadingExceptionOccurred = EventHandlers.ReadingExceptionOccurred,
             });